Cleanup backstack initialization into NavigationStateBridge
- New NavigationStateBridge.kt — a single object owning the whole Activity↔BackstackController seam: the intent escape
handoff, the process-death snapshot (NavStateSnapshot + registry key + save-provider), and one restoreAndPersist(...) holding
the full seeding precedence top-to-bottom.
- Deleted RestoredBackstackTransfer.kt (folded in).
- MainActivity shrank — the restoration ladder, NavStateSnapshot, the registry key, savedStateConfiguration, and 9 unused
imports are gone, replaced by attaching the two task lambdas plus one NavigationStateBridge.restoreAndPersist(...) call.

Behavior is unchanged — same precedence, same config-change no-op, same save provider.
